Move the placeholder into a newly created namespace
